+++ src/3pp/fdt/fdt_sw.c
@@ -72,13 +72,13 @@ static int _fdt_sw_check_header(void *fdt)
 
 static void *_fdt_grab_space(void *fdt, size_t len)
 {
-	int offset = fdt_size_dt_struct(fdt);
+    int offset = fdt_size_dt_struct(fdt);
 	int spaceleft;
 
 	spaceleft = fdt_totalsize(fdt) - fdt_off_dt_struct(fdt)
 		- fdt_size_dt_strings(fdt);
 
-	if ((offset + len < offset) || (offset + len > spaceleft))
+	if ((offset + (int) len < offset) || ((offset + (int) len) > spaceleft))
 		return NULL;
 
 	fdt_set_size_dt_struct(fdt, offset + len);
@@ -89,7 +89,7 @@ int fdt_create(void *buf, int bufsize)
 {
 	void *fdt = buf;
 
-	if (bufsize < sizeof(struct fdt_header))
+	if ((unsigned int)bufsize < sizeof(struct fdt_header))
 		return -FDT_ERR_NOSPACE;
 
 	memset(buf, 0, bufsize);
@@ -117,7 +117,7 @@ int fdt_resize(void *fdt, void *buf, int bufsize)
 	headsize = fdt_off_dt_struct(fdt);
 	tailsize = fdt_size_dt_strings(fdt);
 
-	if ((headsize + tailsize) > bufsize)
+	if ((headsize + tailsize) > (unsigned int) bufsize)
 		return -FDT_ERR_NOSPACE;
 
 	oldtail = (char *)fdt + fdt_totalsize(fdt) - tailsize;
@@ -212,7 +212,7 @@ static int _fdt_find_add_string(void *fdt, const char *s)
 	/* Add it */
 	offset = -strtabsize - len;
 	struct_top = fdt_off_dt_struct(fdt) + fdt_size_dt_struct(fdt);
-	if (fdt_totalsize(fdt) + offset < struct_top)
+	if (fdt_totalsize(fdt) + offset < (unsigned int) struct_top)
 		return 0; /* no more room :( */
 
 	memcpy(strtab + offset, s, len);
